refactor(dev): extract the subagent pump behind one cohesive SubagentView
Review tranche 5 (runner cluster):

- The ~450-line subagent child-stream subsystem moves to subagent-pump.ts:
  run state, pump lifecycle, child-event folding, ghost sweeps, and
  finalize now touch nothing but the client and the SubagentView seam.
  Run creation moves out of the event-translator generator into the
  pump's begin(), so eveEventsToTUIStream no longer owns subagent state.
- SubagentView replaces six individually-optional renderer methods with
  one optional capability whose members are required — a type-legal
  partial implementation (which would ghost placeholders or duplicate
  parent tool rows) can no longer exist. The renderer exposes the view as
  a property over its existing public methods.
- SubagentRun gains its lifecycle authority: status running|settled,
  written once by finalize (now trivially idempotent across the child
  boundary and the subagent.completed fallback), and read by the child
  event fold — a late event on a settled run (a HITL-parked child
  resuming) explicitly reopens it and clears the header's Done mark
  instead of mutating a completed section by accident.

The snapshot-shaped contract (view.render(callId, runSnapshot) replacing
the per-delta methods) is the deliberate remainder: this seam is its
prerequisite, and the redesign deserves its own focused pass.
