+++ packages/eve/src/cli/dev/tui/runner.ts
@@ -19,7 +19,6 @@ import {
   type SubagentCompletedStreamEvent,
   Client,
   ClientSession,
-  isCurrentTurnBoundaryEvent,
 } from "#client/index.js";
 import { loadDevelopmentEnvironmentFiles } from "#cli/dev/environment.js";
 import { subscribeDevelopmentSandboxPrewarmLogs } from "#execution/sandbox/development-prewarm.js";
@@ -28,6 +27,13 @@ import {
   type DevelopmentRuntimeArtifactRefresher,
 } from "#services/dev-client.js";
 import { toErrorMessage } from "#shared/errors.js";
+import { SubagentPump, type SubagentView } from "./subagent-pump.js";
+export type {
+  SubagentRun,
+  SubagentStepUpdate,
+  SubagentToolUpdate,
+  SubagentView,
+} from "./subagent-pump.js";
 import { devBootPhase, type DevBootProgressReporter } from "#internal/dev-boot-progress.js";
 
 import {
@@ -36,7 +42,6 @@ import {
   formatFailureDetail,
   formatFailureHint,
   formatFailureMessage,
-  isAbortLikeError,
   isInterruptedError,
   localFailureHint,
 } from "./errors.js";
@@ -235,31 +240,13 @@ export type AgentTUIRenderer = {
   ): Promise<AgentTUIInputQuestionResponse | undefined>;
   renderStream(result: AgentTUIStreamResult, options?: AgentTUISessionOptions): Promise<void>;
   /**
-   * Out-of-band update for one child step (reasoning + message text) of a
-   * subagent dispatch. Called by the runner as child-session stream events
-   * arrive. The renderer renders this as a body section colored by the
-   * subagent palette.
+   * The renderer's whole subagent surface — sections, nested steps and
+   * tools, ghost sweeps, completion. One optional capability with required
+   * members: a renderer either has a subagent view or it doesn't, and a
+   * type-legal partial implementation (which would ghost placeholders or
+   * duplicate parent tool rows) cannot exist.
    */
-  upsertSubagentStep?(update: SubagentStepUpdate): void;
-  /**
-   * Out-of-band update for one child tool call of a subagent dispatch.
-   */
-  upsertSubagentTool?(update: SubagentToolUpdate): void;
-  /**
-   * Removes one child tool row from a subagent section. Used to drop
-   * `preparing` placeholders whose call never materialized.
-   */
-  removeSubagentTool?(update: { callId: string; childCallId: string }): void;
-  /** Opens a subagent's section the moment its dispatch is known. */
-  beginSubagent?(update: { callId: string; name: string }): void;
-  /** Marks a subagent call complete so its section corner reports Done. */
-  completeSubagent?(update: { callId: string }): void;
-  /**
-   * Registers a tool call id as originating from a subagent's child
-   * session. The renderer must skip or remove parent-level tool blocks for
-   * these ids — they are surfaced via {@link upsertSubagentTool} instead.
-   */
-  markChildToolCallId?(callId: string): void;
+  readonly subagents?: SubagentView;
   /**
    * Out-of-band update for one MCP connection authorization lifecycle.
    * Called by the runner as `authorization.*` events arrive.
@@ -480,13 +467,12 @@ export class EveTUIRunner {
    * Each run holds per-step text accumulators (so reasoning + message land
    * in the same section per child step) and per-tool state.
    */
-  readonly #subagentRuns = new Map<string, SubagentRun>();
+  readonly #subagentPump: SubagentPump;
   /**
    * callId → AbortController for the parallel child-session stream pump
    * launched on `subagent.called`. Cancelled on `subagent.completed`, when
    * the session resets, or when the runner shuts down.
    */
-  readonly #subagentChildPumps = new Map<string, AbortController>();
   /**
    * name → latest known state for one MCP connection
    * authorization lifecycle. Persists across turns because a turn that
@@ -516,6 +502,11 @@ export class EveTUIRunner {
     this.#session = options.session;
     if (options.client !== undefined) this.#client = options.client;
     this.#renderer = createRenderer(options);
+    this.#subagentPump = new SubagentPump({
+      ...(this.#client !== undefined ? { client: this.#client } : {}),
+      ...(this.#renderer.subagents !== undefined ? { view: this.#renderer.subagents } : {}),
+      formatActionResultError,
+    });
     this.#name = options.name ?? "eve";
     this.#tools = options.tools ?? "full";
     this.#reasoning = options.reasoning ?? "full";
@@ -636,7 +627,7 @@ export class EveTUIRunner {
     } finally {
       this.#disposed = true;
       this.#authProbeAbort.abort();
-      this.#abortSubagentChildPumps();
+      this.#subagentPump.abortAll();
       // Restore captured stdout/stderr before a fatal error reaches the CLI.
       this.#unsubscribeDevelopmentSandboxLogs?.();
       this.#unsubscribeDevelopmentSandboxLogs = undefined;
@@ -920,8 +911,7 @@ export class EveTUIRunner {
    * In-flight subagent child-session streams are aborted.
    */
   #startNewSession(): void {
-    this.#abortSubagentChildPumps();
-    this.#subagentRuns.clear();
+    this.#subagentPump.abortAll();
     this.#pendingInputRequests.clear();
     this.#connectionAuthRuns.clear();
     this.#pendingConnectionAuths.clear();
@@ -931,13 +921,6 @@ export class EveTUIRunner {
       this.#session = this.#client.session();
     }
     this.#runtimeArtifacts?.clear();
-  }
-
-  #abortSubagentChildPumps(): void {
-    for (const controller of this.#subagentChildPumps.values()) {
-      controller.abort();
-    }
-    this.#subagentChildPumps.clear();
   }
 
   async #readPromptWithIdleRefresh(options: AgentTUISessionOptions): Promise<string | undefined> {
@@ -1076,10 +1059,9 @@ export class EveTUIRunner {
       events: eveEventsToTUIStream({
         events,
         pendingInputRequests: this.#pendingInputRequests,
-        subagentRuns: this.#subagentRuns,
         turnState,
-        onSubagentCalled: (called) => this.#beginSubagentSection(called),
-        onSubagentCompleted: (callId) => this.#stopSubagentChildPump(callId),
+        onSubagentCalled: (called) => this.#subagentPump.begin(called),
+        onSubagentCompleted: (callId) => this.#subagentPump.settle(callId),
         onConnectionAuthRequired: (event) => this.#handleConnectionAuthRequired(event),
         onConnectionAuthCompleted: (event) => this.#handleConnectionAuthCompleted(event),
         onTerminalFailure: () => {
@@ -1465,342 +1447,6 @@ export class EveTUIRunner {
     if (run.reason !== undefined) update.reason = run.reason;
     this.#renderer.upsertConnectionAuth?.(update);
   }
-
-  /**
-   * The moment a dispatch is known to be a subagent call, its section
-   * header replaces the parent-level tool row (or its still-preparing
-   * placeholder — subagent dispatches never upgrade one, since their
-   * actions are not tool-call kind). Without this the placeholder is
-   * swept at the step boundary and nothing shows until the child's first
-   * content arrives.
-   */
-  #beginSubagentSection(called: SubagentCalledStreamEvent): void {
-    this.#renderer.markChildToolCallId?.(called.data.callId);
-    this.#renderer.beginSubagent?.({ callId: called.data.callId, name: called.data.name });
-    this.#startSubagentChildPump(called);
-  }
-
-  /**
-   * Opens a parallel stream over the child session and folds its events into
-   * nested subagent blocks.
-   *
-   * Pumps are fire-and-forget and must never be awaited at a turn boundary:
-   * a subagent dispatched in `task` mode that parks for HITL never emits a
-   * turn-boundary event on its own stream (`harness/tool-loop.ts` gates
-   * `emitTurnEpilogue` on `mode === "conversation"`), so blocking on a child
-   * stream would stall the prompt until the subagent's serverless function
-   * times out. Pumps stay open across HITL prompts and resume rendering when
-   * the subagent unparks; they end on the child's own boundary or via abort.
-   */
-  #startSubagentChildPump(called: SubagentCalledStreamEvent) {
-    const callId = called.data.callId;
-    if (this.#subagentChildPumps.has(callId)) return;
-    const client = this.#client;
-    if (!client) return;
-
-    const controller = new AbortController();
-    this.#subagentChildPumps.set(callId, controller);
-
-    void (async () => {
-      let boundaryReached = false;
-      try {
-        const childSession = client.session({
-          sessionId: called.data.childSessionId,
-          streamIndex: 0,
-        });
-        const stream = childSession.stream({ signal: controller.signal });
-        for await (const event of stream) {
-          if (controller.signal.aborted) break;
-          this.#applyChildEvent(callId, event);
-          if (isCurrentTurnBoundaryEvent(event)) {
-            // The child's own turn boundary — its final assistant message
-            // is in — is what finishes the section. The parent's
-            // `subagent.completed` arrives independently (often later, once
-            // the parent's turn resumes) and only acts as a fallback.
-            boundaryReached = true;
-            break;
-          }
-        }
-      } catch (error) {
-        if (!isAbortLikeError(error)) {
-          const errorText = toErrorMessage(error);
-          const run = this.#subagentRuns.get(callId);
-          if (run) {
-            const { key, step } = openCurrentSubagentSection(run);
-            step.message = step.message
-              ? `${step.message}\n\nstream error: ${errorText}`
-              : `stream error: ${errorText}`;
-            step.finalized = true;
-            run.currentSectionKey = null;
-            this.#renderer.upsertSubagentStep?.({
-              callId,
-              subagentName: run.name,
-              sectionKey: key,
-              reasoning: step.reasoning,
-              message: step.message,
-              finalized: true,
-            });
-          }
-        }
-      } finally {
-        this.#subagentChildPumps.delete(callId);
-      }
-      if (boundaryReached) this.#finalizeSubagentRun(callId);
-    })();
-  }
-
-  #registerChildTool(
-    callId: string,
-    run: SubagentRun,
-    request: {
-      childCallId: string;
-      toolName: string;
-      input: unknown;
-      status: SubagentToolState["status"];
-    },
-  ): void {
-    const existing = run.tools.get(request.childCallId);
-    const tool: SubagentToolState = existing ?? {
-      toolName: request.toolName,
-      input: request.input,
-      status: request.status,
-    };
-    if (existing) {
-      // Promote status only when the new status is "stronger" — e.g.
-      // approval-requested → executing once the parent approves, but
-      // never demote from done/failed back to executing.
-      const priority: Record<SubagentToolState["status"], number> = {
-        preparing: 0,
-        "approval-requested": 1,
-        executing: 2,
-        done: 3,
-        failed: 3,
-        rejected: 3,
-      };
-      if (priority[request.status] > priority[existing.status]) {
-        existing.status = request.status;
-      }
-      // A late `preparing` announcement must not wipe input the full call
-      // already delivered.
-      if (request.input !== undefined) {
-        existing.input = request.input;
-      }
-    } else {
-      run.tools.set(request.childCallId, tool);
-    }
-    this.#renderer.markChildToolCallId?.(request.childCallId);
-    this.#renderer.upsertSubagentTool?.({
-      callId,
-      subagentName: run.name,
-      childCallId: request.childCallId,
-      toolName: tool.toolName,
-      input: tool.input,
-      status: tool.status,
-    });
-  }
-
-  #stopSubagentChildPump(callId: string) {
-    // Parent reports subagent.completed. The child stream pump terminates
-    // itself on the child's own turn boundary — the authoritative finish
-    // signal, which already finalized the section — so this is a fallback
-    // for runs whose boundary never reached us (a dropped child stream, a
-    // HITL-parked turn resuming later). We do NOT abort the pump here,
-    // because the child's `message.completed` event may still be in flight
-    // (the parent and child streams are independent HTTP connections).
-    this.#finalizeSubagentRun(callId);
-  }
-
-  /**
-   * Settles a subagent section: re-emits a finalized snapshot for any
-   * still-streaming step (flipping its right-title off `streaming`), sweeps
-   * preparing ghosts, and marks the section Done. Idempotent — the child's
-   * turn boundary and the parent's `subagent.completed` can both land here.
-   */
-  #finalizeSubagentRun(callId: string): void {
-    const run = this.#subagentRuns.get(callId);
-    if (!run) return;
-    for (const [sectionKey, step] of run.steps) {
-      if (!step.finalized) {
-        step.finalized = true;
-        this.#renderer.upsertSubagentStep?.({
-          callId,
-          subagentName: run.name,
-          sectionKey,
-          reasoning: step.reasoning,
-          message: step.message,
-          finalized: true,
-        });
-      }
-    }
-    run.currentSectionKey = null;
-    this.#sweepPreparingChildTools(callId, run);
-    this.#renderer.completeSubagent?.({ callId });
-  }
-
-  /**
-   * Drops child tool placeholders that never left `preparing` — their input
-   * never parsed (the child model emitted bad JSON, or the stream ended
-   * mid-generation), so no upgrade is coming and the row would linger as a
-   * `Search …` ghost inside the section.
-   */
-  #sweepPreparingChildTools(callId: string, run: SubagentRun): void {
-    for (const [childCallId, tool] of run.tools) {
-      if (tool.status !== "preparing") continue;
-      run.tools.delete(childCallId);
-      this.#renderer.removeSubagentTool?.({ callId, childCallId });
-    }
-  }
-
-  #applyChildEvent(callId: string, event: HandleMessageStreamEvent) {
-    const run = this.#subagentRuns.get(callId);
-    if (!run) return;
-    const renderer = this.#renderer;
-
-    const emit = (key: number, step: SubagentChildStep) => {
-      renderer.upsertSubagentStep?.({
-        callId,
-        subagentName: run.name,
-        sectionKey: key,
-        reasoning: step.reasoning,
-        message: step.message,
-        finalized: step.finalized,
-      });
-    };
-
-    const finalizeCurrent = () => {
-      if (run.currentSectionKey === null) return;
-      const step = run.steps.get(run.currentSectionKey);
-      if (step) {
-        step.finalized = true;
-        emit(run.currentSectionKey, step);
-      }
-      run.currentSectionKey = null;
-    };
-
-    switch (event.type) {
-      case "reasoning.appended": {
-        const { key, step } = openCurrentSubagentSection(run);
-        step.reasoning = step.reasoning + event.data.reasoningDelta;
-        emit(key, step);
-        break;
-      }
-      case "reasoning.completed":
-        // Reasoning closes within a section but does not close the section
-        // itself — a following `message.appended` should land in the same
-        // box. The section closes on `message.completed` or
-        // `step.completed`.
-        break;
-      case "message.appended": {
-        const { key, step } = openCurrentSubagentSection(run);
-        step.message = step.message + event.data.messageDelta;
-        emit(key, step);
-        break;
-      }
-      case "message.completed": {
-        const { key, step } = openCurrentSubagentSection(run);
-        if (event.data.message !== null && step.message.length === 0) {
-          // Some channels emit only `message.completed` without per-delta
-          // `message.appended` events. Capture the full text in that case.
-          step.message = event.data.message;
-        }
-        step.finalized = true;
-        emit(key, step);
-        run.currentSectionKey = null;
-        break;
-      }
-      case "step.completed":
-        finalizeCurrent();
-        // A valid child call upgrades from `preparing` within its own step;
-        // one still preparing at the boundary never parsed and would linger
-        // as a placeholder ghost in the section.
-        this.#sweepPreparingChildTools(callId, run);
-        break;
-      case "action.preparing": {
-        // The child model committed to a call whose input is still
-        // streaming; show the placeholder row until `actions.requested`
-        // (or `input.requested`) delivers the full input.
-        finalizeCurrent();
-        const data = (event as ActionPreparingStreamEvent).data;
-        this.#registerChildTool(callId, run, {
-          childCallId: data.callId,
-          toolName: data.toolName,
-          input: undefined,
-          status: "preparing",
-        });
-        break;
-      }
-      case "actions.requested": {
-        // Close any pending text section before the tool call so the
-        // tool box renders below it — and the next post-tool message
-        // opens a fresh section.
-        finalizeCurrent();
-        for (const action of event.data.actions) {
-          if (action.kind !== "tool-call") continue;
-          this.#registerChildTool(callId, run, {
-            childCallId: action.callId,
-            toolName: action.toolName,
-            input: action.input,
-            status: "executing",
-          });
-        }
-        break;
-      }
-      case "input.requested": {
-        // Tools that need approval skip `actions.requested` and arrive
-        // here as `input.requested` with the action embedded. Register
-        // the tool section the same way (status: "approval-requested")
-        // so the parent's stale tool box can be suppressed and the
-        // child tool appears under the subagent flow.
-        finalizeCurrent();
-        for (const request of event.data.requests) {
-          if (request.action.kind !== "tool-call") continue;
-          this.#registerChildTool(callId, run, {
-            childCallId: request.action.callId,
-            toolName: request.action.toolName,
-            input: request.action.input,
-            status: "approval-requested",
-          });
-        }
-        break;
-      }
-      case "action.result": {
-        const result = event.data.result;
-        if (result.kind !== "tool-result") break;
-        const tool = run.tools.get(result.callId);
-        if (!tool) break;
-        switch (event.data.status) {
-          case "completed":
-            tool.status = "done";
-            tool.output = result.output;
-            break;
-          case "failed":
-            tool.status = "failed";
-            tool.errorText = formatActionResultError(event);
-            break;
-          case "rejected":
-            tool.status = "rejected";
-            tool.errorText = formatActionResultError(event);
-            break;
-        }
-        const update: SubagentToolUpdate = {
-          callId,
-          subagentName: run.name,
-          childCallId: result.callId,
-          toolName: tool.toolName,
-          input: tool.input,
-          status: tool.status,
-        };
-        if (tool.output !== undefined) update.output = tool.output;
-        if (tool.errorText !== undefined) update.errorText = tool.errorText;
-        renderer.upsertSubagentTool?.(update);
-        break;
-      }
-      default:
-        // Other events (session.*, turn.*, step.started, etc.) carry no
-        // visible text — ignore.
-        break;
-    }
-  }
 }
 
 function createRenderer(options: EveTUIRunnerOptions): AgentTUIRenderer {
@@ -1843,7 +1489,6 @@ function formatAgentUpdateNotice(
 type EveStreamTranslatorInput = {
   events: AsyncIterable<HandleMessageStreamEvent>;
   pendingInputRequests: Map<string, InputRequest>;
-  subagentRuns: Map<string, SubagentRun>;
   turnState: AgentTUITurnState;
   onSubagentCalled?: (event: SubagentCalledStreamEvent) => void;
   onSubagentCompleted?: (callId: string) => void;
@@ -1869,7 +1514,6 @@ async function* eveEventsToTUIStream(
   const {
     events,
     pendingInputRequests,
-    subagentRuns,
     turnState,
     onSubagentCalled,
     onSubagentCompleted,
@@ -2205,21 +1849,9 @@ async function* eveEventsToTUIStream(
         break;
 
       case "subagent.called": {
-        const called = event as SubagentCalledStreamEvent;
-        if (!subagentRuns.has(called.data.callId)) {
-          subagentRuns.set(called.data.callId, {
-            name: called.data.name,
-            steps: new Map(),
-            currentSectionKey: null,
-            nextSectionKey: 0,
-            tools: new Map(),
-          });
-        } else {
-          // Idempotent re-entry (e.g. SSE resume): just refresh the name.
-          const run = subagentRuns.get(called.data.callId);
-          if (run) run.name = called.data.name;
-        }
-        onSubagentCalled?.(called);
+        // Run creation (idempotent for SSE-resume re-entries) lives in the
+        // pump's begin().
+        onSubagentCalled?.(event as SubagentCalledStreamEvent);
         break;
       }
 
@@ -2468,61 +2100,6 @@ function toAgentTUIInputQuestion(request: InputRequest): AgentTUIInputQuestion {
   return question;
 }
 
-type SubagentChildStep = {
-  reasoning: string;
-  message: string;
-  finalized: boolean;
-};
-
-type SubagentToolState = {
-  toolName: string;
-  input: unknown;
-  status: "preparing" | "approval-requested" | "executing" | "done" | "failed" | "rejected";
-  output?: unknown;
-  errorText?: string;
-};
-
-export type SubagentRun = {
-  name: string;
-  /**
-   * One entry per logical "child message" — independent of the child's
-   * `stepIndex` field, which the harness can reuse across multiple
-   * assistant messages within a turn (e.g. a message before a tool call
-   * and another message after the tool result both arrive under
-   * `stepIndex: 0`). The key is a monotonic counter so each
-   * `message.completed` opens a new box on the next inbound delta.
-   */
-  steps: Map<number, SubagentChildStep>;
-  /**
-   * Section currently accepting reasoning/message deltas. `null` means
-   * the next delta opens a new section.
-   */
-  currentSectionKey: number | null;
-  /** Monotonic counter for new section keys. */
-  nextSectionKey: number;
-  tools: Map<string, SubagentToolState>;
-};
-
-export type SubagentStepUpdate = {
-  callId: string;
-  subagentName: string;
-  sectionKey: number;
-  reasoning: string;
-  message: string;
-  finalized: boolean;
-};
-
-export type SubagentToolUpdate = {
-  callId: string;
-  subagentName: string;
-  childCallId: string;
-  toolName: string;
-  input: unknown;
-  status: "preparing" | "approval-requested" | "executing" | "done" | "failed" | "rejected";
-  output?: unknown;
-  errorText?: string;
-};
-
 export type ConnectionAuthChallenge = {
   url?: string;
   userCode?: string;
@@ -2548,18 +2125,3 @@ type ConnectionAuthRun = {
   webhookUrl?: string;
   reason?: string;
 };
-
-function openCurrentSubagentSection(run: SubagentRun): {
-  key: number;
-  step: SubagentChildStep;
-} {
-  if (run.currentSectionKey === null) {
-    run.currentSectionKey = run.nextSectionKey++;
-    run.steps.set(run.currentSectionKey, { reasoning: "", message: "", finalized: false });
-  }
-  const step = run.steps.get(run.currentSectionKey);
-  if (!step) {
-    throw new Error("invariant: subagent section state missing for current key");
-  }
-  return { key: run.currentSectionKey, step };
-}
